+++ ext/js/data/anki-note-builder.js
@@ -18,7 +18,7 @@
 
 import {ExtensionError} from '../core/extension-error.js';
 import {deferPromise} from '../core/utilities.js';
-import {convertHiraganaToKatakana, convertKatakanaToHiragana} from '../language/japanese.js';
+import {convertHiraganaToKatakana, convertKatakanaToHiragana} from '../language/ja/japanese.js';
 import {cloneFieldMarkerPattern, getRootDeckName} from './anki-util.js';
 
 export class AnkiNoteBuilder {
